refactor(tui): split 1626-line module into 6 cohesive submodules
tui.py mixed five visual concerns: color/density constants, pure
runtime-overlay helpers, Textual widgets (cards/lanes/stats/detail/filter),
the modal ticket detail screen + cross-thread message, and the main
KanbanApp + its async wrapper. Splitting along the existing seams:

  tui/
    constants.py    53 — STATE_COLOR, AGENT_COLOR, density flags, lane widths
    helpers.py     197 — _CardStatus + _parse_iso/_silent_seconds/_truncate
                         /_first_meaningful_line/_card_sort_key/_ordered_columns
                         /_compact_rate_limits/_build_runtime_index
                         /_fetch_candidates/_fetch_terminals/_append_token_meta
                         /_safe_id/_matches_filter
    screens.py      81 — _RefreshNow message + TicketDetailScreen modal
    widgets.py     554 — IssueCard / Lane / StatsBar / DetailPane / FilterBar
    app.py         818 — KanbanApp + KanbanTUI (still one unit; key bindings,
                         lifecycle, tracker poll, filter, archive, pause all
                         hang off the same App class so further splitting
                         would force ugly mixins)
    __init__.py     91 — re-exports every name tests / cli / monkeypatch
                         targets rely on

Monkeypatch-safe: `KanbanApp._refresh_tracker` reaches `_fetch_candidates`
through the parent package (`_tui_pkg._fetch_candidates`) so test stubs
applied via `monkeypatch.setattr("symphony.tui._fetch_candidates", ...)`
still intercept the call — a plain local import would bind once at module
load and ignore the patch.

Public API and dotted-path stub targets unchanged. `IssueCard.open_details`
lazy-imports TicketDetailScreen to keep the widgets→screens edge one-way
at module load time. Full pytest suite green: 565 passed, 6 skipped
(identical to baseline).
